Redesign of bypass_service (C++ only)
Python probably won't compile at this point. There are a number of major
changes:

- The complex logic for autoinstantiating and destroying bypass_service
  was ripped out. The user has to manage the lifetime now. For Python,
  lifetime will be managed by with_custodian_and_ward.
- The dedicated thread has been replaced by using an io_service, and
  asio callbacks instead of poll.
- The lock on bypass_service has been removed, avoiding the problems of
  deadlock between this lock and the stream lock. Instead, external
  operations use a run_in_strand function.

There is still the problem of deadlock if a client doesn't service the
ring buffer, which will block up the strand. Even if recv::stream is
modified to drop the lock during push to the ring-buffer, the bypass
service can still be deadlocked, preventing adding new readers. This is
particularly problematic for initialising multiple streams before
servicing any of them.